Refactor outlier_detection_step
Many changes to outlier_detection_step and outlier_detection:
 - add `save_intermediate_files` to user parameters for the step
 - replace `stsci.ImageStats` with `astropy.stats.sigma_clipped_stats`
 - replace `stsci.image.numcombine` with `stsci.image.median`
 - remove `nsigma, hthresh, lthresh, ctegrow` step parameters
 - units of readnoise in flag_cr is now DN, not electrons
