+++ jwst/outlier_detection/outlier_detection.py
@@ -5,17 +5,19 @@ import time
 import numpy as np
 from collections import OrderedDict
 
-from .. import datamodels
-from .. import assign_wcs
-from .. import resample
+from stsci.image import median
+from stsci.tools import bitmask
+from astropy.stats import sigma_clipped_stats
+from scipy import ndimage
 
-from . import flag_cr
-from . import blot_median
-from . import create_median
+from .. import datamodels
+from ..resample import resample, gwcs_blot
 
 import logging
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
+
+CRBIT = np.uint32(datamodels.dqflags.pixel['JUMP_DET'])
 
 
 class OutlierDetection(object):
@@ -40,24 +42,22 @@ class OutlierDetection(object):
 
     """
 
-    def __init__(self, input_models, reffiles=None, to_file=False, **pars):
+    def __init__(self, input_models, reffiles=None, **pars):
         """
         Parameters
         ----------
-        input_models : list of objects or str
+        input_models : list of DataModels, str
             list of data models as ModelContainer or ASN file,
             one data model for each input image
 
         drizzled_models : list of objects
             ModelContainer containing drizzled grouped input images
 
-        to_file : bool
-            Control whether or not to write out intermediate results to files
-
+        reffiles : dict of `jwst.datamodels.DataModel`
+            Dictionary of datamodels.  Keys are reffile_types.
         """
         self.input_models = input_models
         self.reffiles = reffiles
-        self.to_file = to_file
 
         self.num_groups = len(self.input_models.group_names)
 
@@ -107,47 +107,320 @@ class OutlierDetection(object):
             self.outlierpars[kw] = ref_model['outlierpars_table.{0}'.format(kw)]
 
     def do_detection(self):
-        """Compare blotted median to input images to flag outlier pixels in DQ
+        """Flag outlier pixels in DQ of input images
         """
         pars = self.outlierpars
+        save_intermediate_files = pars['save_intermediate_files']
 
         # Start by creating resampled/mosaic images for each group of exposures
-        sdriz = resample.resample.ResampleData(self.input_models, single=True,
+        sdriz = resample.ResampleData(self.input_models, single=True,
             blendheaders=False, **pars)
         sdriz.do_drizzle(**pars)
         drizzled_models = sdriz.output_models
-        if self.to_file:
-            log.info("Saving resampled grouped exposures to disk...")
+        for model in drizzled_models:
+            model.meta.filename += ".fits"
+        if save_intermediate_files:
+            log.info("Writing out resampled exposures...")
             drizzled_models.save()
 
         # Initialize intermediate products used in the outlier detection
         median_model = datamodels.ImageModel(init=drizzled_models[0].data.shape)
-        median_model.meta = drizzled_models[0].meta # provide median with initial metadata
+        median_model.meta = drizzled_models[0].meta
         base_filename = self.input_models[0].meta.filename
-        median_filename = '_'.join(base_filename.split('_')[:2] + ['median.fits'])
-        median_model.meta.filename = median_filename
+        median_model.meta.filename = '_'.join(base_filename.split('_')[:2] +
+            ['median.fits'])
 
         # Perform median combination on set of drizzled mosaics
-        drizzle_groups_sci = [i.data for i in drizzled_models]
-        drizzle_groups_wht = [i.wht for i in drizzled_models]
-        median_model.data = create_median.do_median(drizzle_groups_sci,
-                                        drizzle_groups_wht,
-                                        **pars)
-        if self.to_file:
+        median_model.data = create_median(drizzled_models, **pars)
+
+        if save_intermediate_files:
             log.info("Writing out MEDIAN image to: {}".format(median_model.meta.filename))
             median_model.save(median_model.meta.filename)
+
         # Blot the median image back to recreate each input image specified in
         # the original input list/ASN/ModelContainer
-        blot_models = blot_median.do_blot(median_model, self.input_models,
-                     **pars)
-        if self.to_file:
-            log.info("Writing out BLOT input images...")
+        blot_models = blot_median(median_model, self.input_models, **pars)
+        if save_intermediate_files:
+            log.info("Writing out BLOT images...")
             blot_models.save()
 
         # Perform outlier detection using statistical comparisons between
         # each original input image and its blotted version of the median image
-        flag_cr.do_detection(self.input_models, blot_models,
+        detect_outliers(self.input_models, blot_models,
             self.reffiles, **self.outlierpars)
 
         # clean-up (just to be explicit about being finished with these results)
         del median_model, blot_models
+
+
+def create_median(resampled_models, **pars):
+    """Create a median image from the singly resampled images.
+
+    NOTE: This version is simplified from astrodrizzle's version in the
+        following ways:
+        - type of combination: fixed to 'median'
+        - 'minmed' not implemented as an option
+        - does not use buffers to try to minimize memory usage
+        - astropy.stats.sigma_clipped_stats replaces stsci.imagestats.ImageStats
+        - stsci.image.median replaces stsci.image.numcombine.numCombine
+    """
+    resampled_sci = [i.data for i in resampled_models]
+    resampled_wht = [i.wht for i in resampled_models]
+
+    nlow = pars.get('nlow', 0)
+    nhigh = pars.get('nhigh', 0)
+    maskpt = pars.get('maskpt', 0.7)
+
+    badmasks = []
+    for w in resampled_wht:
+        mean_weight, _, _ = sigma_clipped_stats(w, sigma=3.0, mask_value=0.)
+        weight_threshold = mean_weight * maskpt
+        # Mask pixels were weight falls below MASKPT percent of the mean weight
+        mask = np.less(w, weight_threshold)
+        log.debug("Number of pixels with low weight: {}".format(np.sum(mask)))
+        badmasks.append(mask)
+
+    # Compute median of stack os images using BADMASKS to remove low weight
+    # values
+    median_image = median(resampled_sci, nlow=nlow, nhigh=nhigh,
+        badmasks=badmasks)
+
+    return median_image
+
+
+def blot_median(median_model, input_models, **pars):
+    """Blot resampled median image back to the detector images
+    """
+    interp = pars.get('interp', 'poly5')
+    sinscl = pars.get('sinscl', 1.0)
+
+    # Initialize container for output blot images
+    blot_models = datamodels.ModelContainer()
+
+    log.info("Blotting median...")
+    blot = gwcs_blot.GWCSBlot(median_model)
+
+    for model in input_models:
+        blotted_median = model.copy()
+        blot_root = '_'.join(model.meta.filename.replace('.fits', '').split('_')[:-1])
+        blotted_median.meta.filename = '{}_blot.fits'.format(blot_root)
+
+        # clean out extra data not related to blot result
+        blotted_median.err = None
+        blotted_median.dq = None
+        # apply blot to re-create model.data from median image
+        blotted_median.data = blot.extract_image(model, interp=interp,
+            sinscl=sinscl)
+        blot_models.append(blotted_median)
+
+    return blot_models
+
+
+def detect_outliers(input_models, blot_models, reffiles, **pars):
+    """
+    Flags DQ array for cosmic rays in input images.
+
+    The science frame in each ImageModel in input_models is compared to
+    the corresponding blotted median image in blot_models.  The result is
+    an updated DQ array in each ImageModel in input_models.
+
+    Parameters
+    ----------
+    input_models: JWST ModelContainer object
+        data model container holding science ImageModels, modified in place
+
+    blot_models : JWST ModelContainer object
+        data model container holding ImageModels of the median output frame
+        blotted back to the wcs and frame of the ImageModels in input_models
+
+    reffiles : dict
+        Contains JWST ModelContainers for 'gain' and 'readnoise' reference files
+
+    Returns
+    -------
+    None
+        The dq array in each input model is modified in place
+    """
+
+    gain_models = reffiles['gain']
+    rn_models = reffiles['readnoise']
+
+    for image, blot, gain, rn in zip(input_models, blot_models, gain_models,
+        rn_models):
+        flag_cr(image, blot, gain, rn, **pars)
+
+
+def flag_cr(sci_image, blot_image, gain_image, readnoise_image, **pars):
+    """
+    Masks outliers in science image
+
+    Mask blemishes in dithered data by comparing a science image
+    with a model image and the derivative of the model image.
+
+    Parameters
+    ----------
+    sci_image : ImageModel
+        the science data
+
+    blot_image : ImageModel
+        the blotted median image of the dithered science frames
+
+    gain_image : GainModel
+        the 2-D gain array
+
+    readnoise_image : ReadnoiseModel
+        the 2-D read noise array
+
+    pars : dict
+        the user parameters for Outlier Detection
+
+    Default parameters:
+
+    grow     = 1               # Radius to mask [default=1 for 3x3]
+    ctegrow  = 0               # Length of CTE correction to be applied
+    snr      = "5.0 4.0"       # Signal-to-noise ratio
+    scale    = "1.2 0.7"       # scaling factor applied to the derivative
+    backg    = 0               # Background value
+    """
+
+    grow = pars.get('grow', 1)
+    ctegrow = pars.get('ctegrow', 0) # not provided by outlierpars
+    backg = pars.get('backg', 0)
+    snr1, snr2 = [float(val) for val in pars.get('snr', '5.0 4.0').split()]
+    scl1, scl2 = [float(val) for val in pars.get('scale', '1.2 0.7').split()]
+
+    if sci_image.meta.background.subtracted:
+        subtracted_background = sci_image.meta.background.level
+        log.debug("Subtracted background: {}".format(subtracted_background))
+    else:
+        subtracted_background = backg
+        log.debug("No subtracted background found. "
+            "Using default value from outlierpars: {}".format(backg))
+
+    exptime = sci_image.meta.exposure.exposure_time
+
+    sci_data = sci_image.data * exptime
+    blot_data = blot_image.data * exptime
+    blot_deriv = abs_deriv(blot_data)
+
+    # This mask can take into account any crbits values
+    # specified by the user to be ignored.
+    # dq_mask = build_mask(sci_image.dq, CRBIT)
+
+    gain = gain_image.data
+    rn = readnoise_image.data
+    # TODO: for JWST, the actual readnoise at a given pixel depends on the
+    # number of reads going into that pixel.  So we need to account for that
+    # using the meta.exposure.nints, ngroups and nframes keywords.
+
+    # Define output cosmic ray mask to populate
+    cr_mask = np.zeros(sci_image.shape, dtype=np.uint8)
+
+    ##################   COMPUTATION PART I    ###################
+    # Model the noise and create a CR mask
+    diff_noise = np.abs(sci_data - blot_data)
+    ta = np.sqrt(np.abs(blot_data + subtracted_background) + rn ** 2)
+    t2 = scl1 * blot_deriv + snr1 * ta
+
+    tmp1 = np.logical_not(np.greater(diff_noise, t2))
+
+    # Convolve mask with 3x3 kernel
+    kernel = np.ones((3, 3), dtype=np.uint8)
+    tmp2 = np.zeros(tmp1.shape, dtype=np.int32)
+    ndimage.convolve(tmp1, kernel, output=tmp2, mode='nearest', cval=0)
+
+    ##################   COMPUTATION PART II    ###################
+    # Create a second CR Mask
+    xt2 = scl2 * blot_deriv + snr2 * ta
+
+    np.logical_not(np.greater(diff_noise, xt2) & np.less(tmp2, 9), cr_mask)
+
+    ##################   COMPUTATION PART III    ###################
+    # Flag additional cte 'radial' and 'tail' pixels surrounding CR
+    # pixels as CRs
+
+    # In both the 'radial' and 'length' kernels below, 0=good and
+    # 1=bad, so that upon convolving the kernels with cr_mask, the
+    # convolution output will have low->bad and high->good from which
+    # 2 new arrays are created having 0->bad and 1->good. These 2 new
+    # arrays are then AND'ed to create a new cr_mask.
+
+    # recast cr_mask to int for manipulations below; will recast to
+    # Bool at end
+    cr_mask_orig_bool = cr_mask.copy()
+    cr_mask = cr_mask_orig_bool.astype(np.int8)
+
+    # make radial convolution kernel and convolve it with original cr_mask
+    cr_grow_kernel = np.ones((grow, grow))
+    cr_grow_kernel_conv = cr_mask.copy()
+    ndimage.convolve(cr_mask, cr_grow_kernel, output=cr_grow_kernel_conv)
+
+    # make tail convolution kernel and (shortly) convolve it with original cr_mask
+    cr_ctegrow_kernel = np.zeros((2 * ctegrow + 1, 2 * ctegrow + 1))
+    cr_ctegrow_kernel_conv = cr_mask.copy()
+
+    # which pixels are masked by tail kernel depends on readout direction
+    # We could put useful info in here for CTE masking if needed.  Code
+    # remains below.  For now, we set to zero, which turns off CTE masking.
+    ctedir = 0
+    if (ctedir == 1):
+        cr_ctegrow_kernel[0:ctegrow, ctegrow] = 1
+    if (ctedir == -1):
+        cr_ctegrow_kernel[ctegrow + 1:2 * ctegrow + 1, ctegrow] = 1
+    if (ctedir == 0):
+        pass
+
+    # finally do the tail convolution
+    ndimage.convolve(cr_mask, cr_ctegrow_kernel, output=cr_ctegrow_kernel_conv)
+
+    # select high pixels from both convolution outputs; then 'and' them to
+    # create new cr_mask
+    where_cr_grow_kernel_conv = np.where(cr_grow_kernel_conv < grow * grow, 0, 1)
+    where_cr_ctegrow_kernel_conv = np.where(cr_ctegrow_kernel_conv < ctegrow, 0, 1)
+
+    # combine masks and cast back to Bool
+    np.logical_and(where_cr_ctegrow_kernel_conv, where_cr_grow_kernel_conv, cr_mask)
+    cr_mask = cr_mask.astype(bool)
+
+    count_sci = np.count_nonzero(sci_image.dq)
+    count_cr = np.count_nonzero(cr_mask)
+    log.debug("Pixels in input DQ: {}".format(count_sci))
+    log.debug("Pixels in cr_mask:  {}".format(count_cr))
+
+    # Update the DQ array in the input image in place
+    np.bitwise_or(sci_image.dq, np.invert(cr_mask) * CRBIT, sci_image.dq)
+
+
+def build_mask(dqarr, bitvalue):
+    """ Builds a bit-mask from an input DQ array and a bitvalue flag"""
+
+    bitvalue = bitmask.interpret_bit_flags(bitvalue)
+    if bitvalue is None:
+        return (np.ones(dqarr.shape, dtype=np.uint32))
+    return np.logical_not(np.bitwise_and(dqarr, ~bitvalue)).astype(np.uint32)
+
+
+def abs_deriv(array):
+    """Take the absolute derivate of a numpy array"""
+
+    tmp = np.zeros(array.shape, dtype=np.float64)
+    out = np.zeros(array.shape, dtype=np.float64)
+
+    tmp[1:,:] = array[:-1,:]
+    tmp, out = _absolute_subtract(array, tmp, out)
+    tmp[:-1,:] = array[1:,:]
+    tmp, out = _absolute_subtract(array, tmp, out)
+
+    tmp[:,1:] = array[:,:-1]
+    tmp, out = _absolute_subtract(array, tmp, out)
+    tmp[:,:-1] = array[:,1:]
+    tmp, out = _absolute_subtract(array, tmp, out)
+
+    return out
+
+
+def _absolute_subtract(array, tmp, out):
+    tmp = np.abs(array - tmp)
+    out = np.maximum(tmp, out)
+    tmp = tmp * 0.
+    return tmp, out
